move useDeleteShare()
