+++ projects/graphql-client/src/index.ts
@@ -30,3 +30,4 @@ export * from "./remote/mutations/useRenamePlaylist"
 export * from "./remote/mutations/useRemoveSongsFromPlaylist"
 export * from "./remote/mutations/useRemoveSongFromLibrary"
 export * from "./remote/mutations/useLeaveShare"
+export * from "./remote/mutations/useDeleteShare"
